requirements.txt
tensorflow==2.13.0
fastapi
uvicorn
numpy
pillow
scikit-learn
opencv-python